refactor: 既存テストを分割 (#9)
